requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.7.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
dataclasses==0.8
Django==3.1.5
django-baton==2.1.5
djangorestframework==3.12.2
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-cloud-vision==2.7.3
googleapis-common-protos==1.56.3
greenlet==1.1.0
grpcio==1.48.2
grpcio-status==1.48.2
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
importlib-metadata==4.6.1
isort==5.7.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.1.1
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.19.5
packaging==20.9
pandas==1.1.5
pep8==1.7.1
Pillow==8.4.0
proto-plus==1.22.1
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2020.5
requests==2.25.1
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==3.1.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
